requirements.txt
django<3.3
django-admin-sortable2==1.0.2
environs[django]==9.3.4
Pillow==8.4.0
django-tinymce==3.3.0